+++ src/main/java/com/example/crrr/repository/ClientsRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ClientsRepository extends JpaRepository<Clients, Integer> { }
+public interface ClientsRepository extends JpaRepository<Clients, Long> { }
