+++ src/main/java/com/example/crrr/repository/TeacherRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface TeacherRepository extends JpaRepository<Teacher, Integer> {
+public interface TeacherRepository extends JpaRepository<Teacher, Long> {
     @EntityGraph(value = "teacher-with-groups-and-course")
     @Query("SELECT t FROM Teacher t")
     List<Teacher> findAllWithGroupsAndCourse();
